feat(#28) :: 빵 유형 설문 2/3/5단계 — 제출·결과 저장, v2 API, AI 시즌 연동
luvin-survey-change-plan/BACKEND_CHANGES.md 구현 순서 2~5단계(프론트 연동인
4단계 제외)를 반영. 1단계(순수 채점/분류 로직, c8e66a1)는 이미 반영되어 있음.

- Flyway 도입(spring-boot-flyway — Boot 4에서 자동설정이 별도 모듈로 분리된 것을
  실제 부팅으로 확인). 기존 34개 테이블은 ddl-auto:update가 계속 관리하고,
  survey_definitions/submissions/submission_answers/results,
  ai_season_creation_inputs 5개 신규 테이블만 V1 migration으로 추가
- SurveySubmissionServiceImpl: member row lock으로 동시 제출 직렬화,
  (memberId, clientSubmissionId) unique + canonical payload hash로 재시도
  멱등성/충돌(409) 판정, 제출+20답변+결과 저장과 User.latestSurveyResultId/
  personalityType 갱신을 하나의 transaction으로 묶음
- 신규 API: GET /api/surveys/v2/definition, POST .../submissions,
  GET .../results/me, GET .../submissions/{clientSubmissionId}.
  SurveyV2ExceptionHandler로 공통 계약의 에러코드(400/404/409/410/503) 구현,
  ApiResponse에 nullable code 필드 추가(기존 응답 호환). 기존
  /api/surveys/{questionId}, /api/surveys/submit은 변경 없음
- User.surveyCompleted 판정 기준을 existsByMemberId(답변 1개만 있어도 true)에서
  latestSurveyResultId 연결 여부로 변경
- AiTraitsRequest/TraitsDto를 Integer -> BigDecimal로 전환. AI가 실제로 decimal을
  받는지 확인되지 않아 AiTraitsWireAdapter가 app.ai.decimal-traits-enabled
  (기본 false)에 따라 wire 직전에만 명시적으로 정수 반올림
- POST /api/ai/seasons/from-survey: 서버가 저장한 설문 결과(gender + 13개 core
  점수)로 시즌 생성. 원격 AI 호출 전에 ai_season_creation_inputs에 입력을
  선저장해 재설문 경합에도 retry payload를 고정. AiSeasonStatusView에
  sourceSurveyResultId 추가(legacy 시즌은 null)
- 버그 수정 2건: (1) Flyway가 Hibernate ddl-auto보다 먼저 실행되어 빈 DB에서
  users/ai_seasons FK가 실패하던 것을 엔티티 필드 방식으로 변경,
  (2) User.gender(대문자 "FEMALE")를 AI 계약(소문자 male/female)으로 변환하지
  않고 있던 것 수정
- openapi.json 재생성

임시 통합 테스트로 확인 후 삭제: 제출 전체 플로우(정의 조회→제출→멱등 재시도→
충돌→조회→타 유저 격리), from-survey 시즌 생성(gender/traits 변환, AI 멱등
재호출 방지), 21개 답변·타 문항 answerId·retired definition 재제출·동시 제출
경합(6 스레드가 result 1개로 수렴, submission row 중복 없음). 전체
./gradlew clean test 실 Postgres 회귀 통과.

+++ src/main/java/com/luvin/ai/client/dto/TraitsDto.java
@@ -1,22 +1,29 @@
 package com.luvin.ai.client.dto;
 
+import java.math.BigDecimal;
+
 /**
  * 13개 성향 점수. openapi.json의 Traits 스키마와 1:1 대응.
  * 모든 필드 필수, 1..100 범위. Spring이 AI 호출 전에 검증한다 (요구사항 3.3).
+ *
+ * BigDecimal인 이유: 설문 v2 채점은 소수 둘째 자리까지 의미 있는 점수(예 75.25)를 만든다.
+ * AI 서비스가 실제로 decimal을 받는지 계약이 확인되기 전까지는
+ * {@link com.luvin.ai.service.AiTraitsWireAdapter}가 wire 직전에 명시적으로 정수 반올림한다 —
+ * 이 타입 자체는 정밀도를 그대로 유지해 향후 decimal 계약이 확정되면 절삭 없이 바로 전환할 수 있다.
  */
 public record TraitsDto(
-        Integer affectionExpression,
-        Integer relationshipAnxiety,
-        Integer relationshipAvoidance,
-        Integer emotionalAttunement,
-        Integer relationshipInitiative,
-        Integer practicalPriority,
-        Integer reassuranceNeed,
-        Integer jealousyReactivity,
-        Integer relationshipEnergyDependence,
-        Integer emotionalSuppression,
-        Integer conflictConfrontation,
-        Integer relationshipPace,
-        Integer interestExpressionFrequency
+        BigDecimal affectionExpression,
+        BigDecimal relationshipAnxiety,
+        BigDecimal relationshipAvoidance,
+        BigDecimal emotionalAttunement,
+        BigDecimal relationshipInitiative,
+        BigDecimal practicalPriority,
+        BigDecimal reassuranceNeed,
+        BigDecimal jealousyReactivity,
+        BigDecimal relationshipEnergyDependence,
+        BigDecimal emotionalSuppression,
+        BigDecimal conflictConfrontation,
+        BigDecimal relationshipPace,
+        BigDecimal interestExpressionFrequency
 ) {
 }
